ultimas implemtações

+++ app/src/main/java/com/jessemanarim/entregaexpressa/feature_entrega/presentation/delivery_list/DeliveriesListFragment.kt
@@ -1,5 +1,6 @@
 package com.jessemanarim.entregaexpressa.feature_entrega.presentation.delivery_list
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -17,12 +18,12 @@ import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-
 class DeliveriesListFragment : Fragment() {
 
     private var _binding: FragmentDeliveryListBinding? = null
     private val binding get() = _binding!!
     private lateinit var adapter: DeliveriesAdapter
+    private var _deliveries = emptyList<Delivery>()
 
     private val _viewModel: DeliveryListViewModel by viewModel()
 
@@ -32,6 +33,23 @@ class DeliveriesListFragment : Fragment() {
     ): View {
         _binding = FragmentDeliveryListBinding.inflate(inflater, container, false)
 
+        subscribeStateFlow()
+
+        return binding.root
+    }
+
+    private fun prepareAdapter(){
+        adapter = DeliveriesAdapter(
+            _deliveries,
+            deleteClick = { currentDelivery -> confirmRegisterExclusion(currentDelivery) },
+            openDetailClick = { currentDelivery -> navigateToDetails(currentDelivery) }
+        )
+
+        binding.recyclerView.adapter = adapter
+        binding.recyclerView.layoutManager = LinearLayoutManager(context)
+    }
+
+    private fun subscribeStateFlow(){
         lifecycleScope.launchWhenStarted {
             _viewModel.uiState.collectLatest { uiState ->
                 when(uiState){
@@ -47,13 +65,10 @@ class DeliveriesListFragment : Fragment() {
                 }
             }
         }
-
-        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
         binding.fab.setOnClickListener {
             findNavController().navigate(R.id.action_DeliveryListFragment_to_DeliveryCreationFragment)
         }
@@ -65,21 +80,25 @@ class DeliveriesListFragment : Fragment() {
     }
 
     private fun loadDeliveryList(list: List<Delivery>){
-        adapter = DeliveriesAdapter(
-            list,
-            deleteClick = { currentDelivery ->
+        _deliveries = list
+        prepareAdapter()
+    }
+
+    fun navigateToDetails(currentDelivery: Delivery){
+        val bundle = Bundle()
+        bundle.putInt("deliveryId", currentDelivery.deliveryId)
+        findNavController().navigate(R.id.action_DeliveryListFragment_to_DeliveryDetailFragment, bundle)
+    }
+
+    private fun confirmRegisterExclusion(currentDelivery: Delivery){
+        AlertDialog.Builder(requireContext())
+            .setMessage(getString(R.string.confirm_exclusion_message))
+            .setPositiveButton(getString(R.string.yes)) { dialog, which ->
                 CoroutineScope(Dispatchers.IO).launch {
                     _viewModel.deleteDelivery(currentDelivery)
                 }
-            },
-            openDetailClick = { currentDelivery ->
-                val bundle = Bundle()
-                bundle.putInt("deliveryId", currentDelivery.deliveryId)
-                findNavController().navigate(R.id.action_DeliveryListFragment_to_DeliveryDetailFragment, bundle)
             }
-        )
-
-        binding.recyclerView.adapter = adapter
-        binding.recyclerView.layoutManager = LinearLayoutManager(context)
+            .setNegativeButton(getString(R.string.no), null)
+            .show()
     }
 }
